Catch Import error if playwright is not installed so that users can see proper error messages and playwright is not a required library
